requirements.txt
crewai>=0.70.0
langchain-groq>=0.1.0
groq>=0.1.0
python-dotenv>=1.0.0